discard string context of git config's core.excludesFile
Before this change, any store path in your git config would become part
of the string context of any of the parsed strings, and then prevent us
from appending it to a path.

Upon further reflection, we don't care even if core.excludesFile is a
store path itself, so we can unconditionally throw away the string
context.